Move Result into GitHub from GitHub::KV
